fix!: Fix revisions of `openPage` functions and add incompatibility notice
I haven't found a way to reliably scroll smoothly to hashes on the same page in Chrome when using `scroll-behavior: smooth` on `:root`.

BREAKING CHANGE: I haven't found a way to get `openPage` to work reliably in combination with `scroll-behavior: smooth` in Chrome. This fixes problems in the attempts to fix the issue, but also concludes by stating an incompatibility with `scroll-behavior: smooth`. ALL PRIOR VERSIONS MAY BE CONSIDERED DEPRECATED.